Extract Postgres version helpers
