+++ db/main/migrate/20180823000000_add_preferences.rb
@@ -1,16 +1,9 @@
 class AddPreferences < ActiveRecord::Migration[4.2]
+  include Travis::PostgresVersion
+
   def change
     change_table :users do |t|
       t.column :preferences, json_type
     end
-  end
-
-  def json_type
-    postgres_version == '9.3' ? :json : :jsonb
-  end
-
-  def postgres_version
-    full = ActiveRecord::Base.connection.select_value('SELECT version()')
-    full[/^PostgreSQL (\d+\.\d+)/, 1]
   end
 end
