+++ db/main/migrate/20190102000000_add_organization_preferences.rb
@@ -1,16 +1,9 @@
 class AddOrganizationPreferences < ActiveRecord::Migration[5.2]
+  include Travis::PostgresVersion
+
   def change
     change_table :organizations do |t|
       t.column :preferences, json_type
     end
-  end
-
-  def json_type
-    postgres_version == '9.3' ? :json : :jsonb
-  end
-
-  def postgres_version
-    full = ActiveRecord::Base.connection.select_value('SELECT version()')
-    full[/^PostgreSQL (\d+\.\d+)/, 1]
   end
 end
